+++ main.js
@@ -1,59 +1,54 @@
+const {askQuestion} = require("./utils.js");
+
+
+const createScores = (players) => new Array(players.length).fill(501) 
+
 const sumStr = (elem) => (elem === 'DB')? 50: 25;
+
 const whatToDo = (elem) => (typeof elem === 'string') ? sumStr(elem) : elem.reduce((x, y) => (x * y));
 
-const Player = (name) => {
-    let score = 501;
-    return {
-      play (array) { 
-          var x = array.map(element => whatToDo(element));
-          score -= x.reduce((x, y) => (x + y));
-          score = Math.abs(score);
-          console.log(name, score)
-      },
-      getName(){
-          return name;
-      },
-      getScore(){
-          return score;
-      }
-    }
- }
 
-const initGame = (names) => {
-    return names.map(name => Player(name));
+const play = async (score, player) => {
+    console.log("Ingrese jugada de ", player) 
+    let jugadas = await askQuestion("-->: ");
+    jugadas = JSON.parse(jugadas)
+    const points = jugadas.map(element => whatToDo(element)).reduce((x, y) => (x + y));
+    return Math.abs(score - points)
 }
 
-const game = (...names) => {
-    let plays = [['DB', [3,20], [3,19]], ['SB', [2,20], [3,20]], [[1, 18], [1, 10], [1, 12]]];
-    console.log(names)
-    var players = initGame(names);
-    var currentPlayer = 0;
-    return {
-        playGame() {
-            var curr = currentPlayer;
-            currentPlayer = currentPlayer + 1 === players.length ? 0 : currentPlayer + 1;
-            var play = plays.shift();
-            var jugada = play === undefined?'return':play;
-            if (jugada === 'return') {
-                return
-            }
-            players[curr].play(jugada) == 0?endGame():this.playGame();
-        },
-        endGame() {
-            console.log('terminado')
-        }
-    }
+const new_scores = (i_player, scores, score_2) => {
+    return scores.slice(0,i_player).concat([score_2]).concat(scores.slice(i_player+1))
 }
 
-// const p1 = Player('p1')
-// p1.play(['SB', [2,20], [3,20]])
-// console.log(p1.getName())
-// console.log(p1.getScore())
+const playGame = async (players, scores, i_players, i_plays) => {
+    const i_players_2 = i_players + 1 === players.length ? 0 : i_players + 1;
+    const i_plays_2 = i_plays + 1;
+    const score_2 = await play(scores[i_players], players[i_players]);
+    console.log(score_2)
+    const scores_2 = new_scores(i_players, scores, score_2);
+    (score_2 == 0) ? endGame(players, scores_2) : await playGame(players, scores_2, i_players_2, i_plays_2);
+}
 
-const g = game('a','b')
-g.playGame()
-// const noFinish = (players) => (players.some((player) => player.getScore() === 0))?'juego termiando': noFinish; 
+const logResults = (players, scores, i) => {
+    console.log(players[i], scores[i])
+    players.length == i + 1 ? {} : logResults(players, scores, i + 1);
+}
 
-// const didWin = (player) => player.getScore() === 0 ? true : false
-// const didFinish = (players) => players.reduce(player => didWin(player) && true );
-// const play = (game, movements) => didFinish(game.getPlayers()) ? true : game.getPlayers().map(player => player.play)
+const endGame = (players, scores) => {
+    const winner = scores.findIndex((elem) => elem === 0);
+    const why = winner === -1 ? 'no hay más jugadas' : 'el ganador es '.concat(players[winner])
+    console.log('Juego terminado porque', why)
+    logResults(players, scores, 0)
+}
+
+const getPlayers = async () => {
+    const players = await askQuestion("Ingrese nombres de jugadores separados por comas: ")
+    return players.split(",")
+}
+
+
+(async () => {
+    const players = await getPlayers()
+    const scores = createScores(players)
+    await playGame(players, scores, 0, 0);
+})()
